moved fn box_cxcywh_to_xyxy to anot_utils from dino
